requirements.txt
aiohappyeyeballs==2.6.1
aiohttp>=3.8.5
aiosignal==1.3.2
async-timeout==5.0.1
attrs==25.3.0
certifi==2025.1.31
charset-normalizer==3.4.1
discord.py>=2.3.0
dnspython==2.7.0
frozenlist==1.5.0
idna==3.10
multidict==6.4.2
orjson==3.10.16
pomice>=1.2.0
propcache==0.3.1
psutil==7.0.0
pymongo>=4.5.0
python-dotenv>=1.0.0
redis==5.2.1
requests==2.32.3
spotipy>=2.23.0
typing_extensions==4.13.2
urllib3==2.4.0
wavelink>=2.6.3
websockets==15.0.1
yarl==1.19.0
asyncio>=3.4.3